LayoutTests/fast/events/tabindex-no-focusable-expected.txt

PASS document.activeElement.id is "MoveToMe"
PASS document.activeElement.id is "MoveToMe"
PASS successfullyParsed is true

TEST COMPLETE
